address pr comments - refactor to create saved search as an api call
