+++ cypress/utils/commands.js
@@ -302,6 +302,8 @@ Cypress.Commands.add('deleteWorkspace', (workspaceName) => {
 });
 
 Cypress.Commands.add('createInitialWorkspaceWithDataSource', (dataSourceTitle, workspaceName) => {
+  cy.intercept('POST', '/api/workspaces').as('createWorkspaceInterception');
+
   cy.getElementByTestId('workspace-initial-card-createWorkspace-button')
     .should('be.visible')
     .click();
@@ -320,6 +322,11 @@ Cypress.Commands.add('createInitialWorkspaceWithDataSource', (dataSourceTitle, w
     .trigger('click');
   cy.getElementByTestId('workspace-detail-dataSources-associateModal-save-button').click();
   cy.getElementByTestId('workspaceForm-bottomBar-createButton').should('be.visible').click();
+
+  cy.wait('@createWorkspaceInterception').then((interception) => {
+    // save the created workspace ID as an alias
+    cy.wrap(interception.response.body.result.id).as('WORKSPACE_ID');
+  });
   cy.contains(/successfully/);
 });
